Fix the issue that causes crash of generated ld.so for ami loader
In amazon Linux C5 image, the transformed loader crashes due to reading
uninitialized data after .bss. This is ok in original loader, since the memory
after .bss should be zero. However, in our case, we insert an extra segment,
the file data between .bss and the 1st section of inserted segment remains.
This area is not fully empty. So we clean it.